Added FinanceAbility
This ability allows Jarvis to help with managing expenses in a shared
living environment where partners split the costs in half.

At the end of each month or at any point, members in the household can
request an expense report from Jarvis, letting members know who has
paid more at that given time point.

This makes it easy to even out expenses and for the parties to compensate
whoever has paid more than the other for that given period.